+++ src/IbanNet.FluentValidation/Properties/AssemblyInfoAttributes.cs
@@ -1,4 +1,6 @@
 ﻿using System.Resources;
+using System.Runtime.CompilerServices;
 
+[assembly: InternalsVisibleTo("IbanNet.FluentValidation.Tests")]
 [assembly: NeutralResourcesLanguage("en")]
 [assembly: CLSCompliant(true)]