+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v5.9.0
 
+- Upgraded SWIFT registry to April '23 release 94, which adds Mongolia (MN) and Nicaragua (NI). Importantly, the spec seems to differ from what is listed on Wikipedia (28 vs 32 char length).
+  Note: the last updated timestamp of Georgia (GE) was also changed, because of an update to contact information in the spec (unused by IbanNet).
 - Refactored pattern tokenizer to support wide tokens (multi char), next to (single) char tokens. This is an internal change that can unlock some extra future functionality.
 
 ## v5.8.1